histoqc._pipeline: fix worker issues in multiprocessing scenario
In some interpreters (ie IPython) multiprocessing will be unable
to access the worker functions. This fixes the issue.

Also improves robustness of cleanup under user requested abort
